moved rejection request into separate component; upon delete changed history to datetime.now
