+++ PuzzlePost/client/src/components/Puzzle/PuzzleRejection.js
@@ -1,0 +1,57 @@
+import React, { useContext, useState } from "react";
+import { Button, Modal, ModalHeader, ModalBody, ModalFooter, Form, FormGroup, Label, Input } from "reactstrap";
+import { RequestContext } from "../../providers/RequestProvider";
+
+
+const PuzzleRejection = ({ modal, toggle, request }) => {
+    const { postRejection } = useContext(RequestContext);
+    const [rejection, setRejection] = useState({
+        puzzleId: request.puzzleId,
+        requestingPuzzleUserId: request.requestingPuzzleUserId,
+        content: ""
+    })
+
+    const handleFieldChange = (e) => {
+        const stateToChange = { ...rejection };
+        stateToChange[e.target.id] = e.target.value;
+        setRejection(stateToChange);
+    };
+
+    const rejectRequest = () => {
+        postRejection(rejection);
+        toggle();
+        // getAllPendingRequests(parseInt(activeUser.id));
+    }
+
+    return (
+        <div>
+
+            <Modal isOpen={modal} toggle={toggle} className="rejection">
+                <ModalHeader toggle={toggle}>Enter a Reason (if you wish)</ModalHeader>
+                <ModalBody>
+                    <Form className="rejectionForm">
+                        <FormGroup>
+                            <Label className="rejectionContentLabel">Message:</Label>
+                            <Input
+                                className="rejectionContent"
+                                onChange={handleFieldChange}
+                                type="textarea"
+                                id="content"
+                                value={rejection.content}
+                                placeholder="Enter Your Rejection Message Here"
+                            />
+                        </FormGroup>
+                    </Form>
+                </ModalBody>
+                <ModalFooter>
+                    <Button color="primary" onClick={rejectRequest}>Send Rejection</Button>{' '}
+                    <Button color="secondary" onClick={toggle}>Just Kidding</Button>
+                </ModalFooter>
+            </Modal>
+        </div>
+
+    )
+}
+
+export default PuzzleRejection;
+
